Automacao_Tarefas/requirements.txt
pdf2docx
python-docx
pypdf